Modify ToC retrieval to use page scraping
Salesforce has done away with the publically accessible xml Table of
Contents files we previously utilised to work out what articles to
present and where to link to.

Thankfully, the html on page is programmatically structured (see
https://github.com/forcedotcom/doc-portal/blob/0d46efa2b5b0f178879e00a90594f7a294d9a57a/helpers/create_toc.xsl
by way of illustration - I believe this is the project used for
generating the docs. To further back that up, it appears that the Table
of Contents is not publically exposed, being retrieved by a database
call in routing in a `get_toc` server-side method in https://github.com/forcedotcom/doc-portal/blob/0d46efa2b5b0f178879e00a90594f7a294d9a57a/helpers/helpers.rb)

As such, we can programmatically deconstruct it, unfortunately we just
have to retrieve all of the cruft that comes with an html page intended
for use in by a user in a browser (instead of a neat XML file intended
for programmatic consumption)