Prevent dups on manual retrieve during start cache
Fixes Issue #16

If a user fired off a retrieve from the command palette during Sublime
startup caching (for a doc type that was being cached), you would wind
up with duplicates.

To fix this, the cache now implements MutableSet as well as
MutableSequence, and existing methods check for existence of duplicates.
The whole structure could use some optimisation, but will get us over
the line for now