Write new JSON ToC based retrievals
There's been another update to the SF doc portal - it appears the doc
portal's front-end code has been revamped pretty thoroughly in
everything but visual appearance of the page, being replaced with an
angular app.

This has broken the scraping method, as the contents of the page are
now loaded in using Javascript. Scraping could in theory be made to
work again, but it would require a headless browser to render the page
which would be a real challenge from inside Sublime.

After a bit of digging in the (minified!) source for the SF Doc angular
app, I turned up some code that would populate the in-page Table of
Contents (ToC) from a URL structure that turned out to serve up a JSON
ToC, depending on what "document type" you plugged in. I've used this
to create new documentation retrieval strategies